+++ internal/codebase/codebase_test.go
@@ -10,6 +10,8 @@ import (
 	"os"
 	"path/filepath"
 	"reflect"
+	"strings"
+	"sync"
 	"testing"
 )
 
@@ -17,7 +19,7 @@ func TestCodebase_Projects(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
 
-	projects := map[string]manifest.Project{
+	expectedProjects := map[string]manifest.Project{
 		"test/15": {Remote: "test"},
 	}
 
@@ -25,7 +27,7 @@ func TestCodebase_Projects(t *testing.T) {
 	manProviderMock.EXPECT().
 		Read("test-dir/.srcode/manifest.json").
 		Return(manifest.Manifest{
-			Projects: projects,
+			Projects: expectedProjects,
 		}, nil)
 
 	codebase := &codebase{
@@ -38,7 +40,7 @@ func TestCodebase_Projects(t *testing.T) {
 		t.FailNow()
 	}
 
-	if !reflect.DeepEqual(projects, projects) {
+	if !reflect.DeepEqual(projects, expectedProjects) {
 		t.FailNow()
 	}
 }
@@ -252,21 +254,117 @@ func TestCodebase_Sync(t *testing.T) {
 	repoProviderMock.EXPECT().Open(filepath.Join(dir, "test/c/d")).Return(cRepoMock, nil)
 	cRepoMock.EXPECT().SetConfig("user.mail", "alois@micard.lu").Return(nil)
 
-	added, deleted, err := codebase.Sync(true)
-	if err != nil {
+	wg := sync.WaitGroup{}
+
+	added := map[string]manifest.Project{}
+	addedChan := make(chan ProjectEntry)
+	go func() {
+		wg.Add(1)
+		for entry := range addedChan {
+			added[entry.Path] = entry.Project
+		}
+
+		wg.Done()
+	}()
+
+	deleted := map[string]manifest.Project{}
+	deletedChan := make(chan ProjectEntry)
+	go func() {
+		wg.Add(1)
+		for entry := range deletedChan {
+			deleted[entry.Path] = entry.Project
+		}
+
+		wg.Done()
+	}()
+
+	if err := codebase.Sync(true, addedChan, deletedChan); err != nil {
 		t.FailNow()
 	}
 
-	if len(added) != 1 {
+	wg.Wait()
+
+	if len(added) != 1 || added["test-12"].Remote != "test.git" {
 		t.Fail()
 	}
-	if len(deleted) != 1 {
+	if len(deleted) != 1 || deleted["test/a/b"].Remote != "test.git" {
 		t.Fail()
 	}
 
 	// make sure we have deleted the directory
 	if _, err := os.Stat(filepath.Join(dir, "test", "a", "b")); !os.IsNotExist(err) {
 		t.Fail()
+	}
+}
+
+func TestCodebase_Sync_NoChannel(t *testing.T) {
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+
+	repoProviderMock := repository_mock.NewMockProvider(mockCtrl)
+	manProviderMock := manifest_mock.NewMockProvider(mockCtrl)
+	repoMock := repository_mock.NewMockRepository(mockCtrl)
+
+	dir := t.TempDir()
+
+	codebase := &codebase{
+		repoProvider: repoProviderMock,
+		repo:         repoMock,
+		manProvider:  manProviderMock,
+		rootPath:     dir,
+	}
+
+	// create mock directory
+	if err := os.MkdirAll(filepath.Join(dir, "test", "a", "b"), 0750); err != nil {
+		t.FailNow()
+	}
+
+	manProviderMock.EXPECT().
+		Read(filepath.Join(dir, metaDir, manifestFile)).
+		Return(manifest.Manifest{
+			Projects: map[string]manifest.Project{
+				"test/a/b": {Remote: "test.git"},
+				"test/c/d": {Remote: "test.git"},
+			},
+		}, nil)
+
+	repoMock.EXPECT().Pull("origin", "master").Return(nil)
+	repoMock.EXPECT().Push("origin", "master").Return(nil)
+
+	manProviderMock.EXPECT().
+		Read(filepath.Join(dir, metaDir, manifestFile)).
+		Return(manifest.Manifest{
+			Projects: map[string]manifest.Project{
+				"test-12": {
+					Remote: "test.git",
+					Config: map[string]string{
+						"user.name": "Aloïs Micard",
+					},
+				},
+				"test/c/d": {
+					Remote: "test.git",
+					Config: map[string]string{
+						"user.mail": "alois@micard.lu",
+					},
+				},
+			},
+		}, nil)
+
+	// should clone missing projects
+	repoProviderMock.EXPECT().
+		Clone("test.git", filepath.Join(dir, "test-12")).
+		Return(nil, nil)
+
+	cRepoMock := repository_mock.NewMockRepository(mockCtrl)
+	repoProviderMock.EXPECT().Open(filepath.Join(dir, "test-12")).Return(cRepoMock, nil)
+	cRepoMock.EXPECT().SetConfig("user.name", "Aloïs Micard").Return(nil) // restore config
+
+	cRepoMock = repository_mock.NewMockRepository(mockCtrl)
+	repoProviderMock.EXPECT().Open(filepath.Join(dir, "test/c/d")).Return(cRepoMock, nil)
+	cRepoMock.EXPECT().SetConfig("user.mail", "alois@micard.lu").Return(nil)
+
+	if err := codebase.Sync(true, nil, nil); err != nil {
+		t.FailNow()
 	}
 }
 
@@ -325,6 +423,107 @@ func TestCodebase_Run(t *testing.T) {
 
 	// Try to run a global command
 	if res, err := codebase.Run("greet-global"); err != nil || res != "Hello from global command" {
+		t.Fail()
+	}
+}
+
+func TestCodebase_BulkGIT(t *testing.T) {
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+
+	manProviderMock := manifest_mock.NewMockProvider(mockCtrl)
+	repoProviderMock := repository_mock.NewMockProvider(mockCtrl)
+
+	codebase := &codebase{
+		manProvider:  manProviderMock,
+		repoProvider: repoProviderMock,
+		rootPath:     "/etc/code",
+		localPath:    "a",
+	}
+
+	manProviderMock.EXPECT().
+		Read(filepath.Join("/", "etc", "code", metaDir, manifestFile)).
+		Return(manifest.Manifest{
+			Projects: map[string]manifest.Project{
+				"test/something-1": {},
+				"test/something-2": {},
+			},
+		}, nil)
+
+	for _, path := range []string{"test/something-1", "test/something-2"} {
+		repoMock := repository_mock.NewMockRepository(mockCtrl)
+		repoProviderMock.EXPECT().
+			Open(filepath.Join("/", "etc", "code", path)).
+			Return(repoMock, nil)
+
+		repoMock.EXPECT().
+			RawCmd([]string{"pull", "--rebase"}).
+			Return("", nil)
+	}
+
+	if err := codebase.BulkGIT([]string{"pull", "--rebase"}, nil); err != nil {
+		t.Fail()
+	}
+}
+
+func TestCodebase_BulkGIT_WithOut(t *testing.T) {
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+
+	manProviderMock := manifest_mock.NewMockProvider(mockCtrl)
+	repoProviderMock := repository_mock.NewMockProvider(mockCtrl)
+
+	codebase := &codebase{
+		manProvider:  manProviderMock,
+		repoProvider: repoProviderMock,
+		rootPath:     "/etc/code",
+		localPath:    "a",
+	}
+
+	manProviderMock.EXPECT().
+		Read(filepath.Join("/", "etc", "code", metaDir, manifestFile)).
+		Return(manifest.Manifest{
+			Projects: map[string]manifest.Project{
+				"test/something-1": {},
+				"test/something-2": {},
+			},
+		}, nil)
+
+	wg := sync.WaitGroup{}
+
+	sb := strings.Builder{}
+	ch := make(chan string)
+	go func() {
+		wg.Add(1)
+		for res := range ch {
+			sb.WriteString(res)
+			sb.WriteString("\n")
+		}
+		wg.Done()
+	}()
+
+	for _, path := range []string{"test/something-1", "test/something-2"} {
+		repoMock := repository_mock.NewMockRepository(mockCtrl)
+		repoProviderMock.EXPECT().
+			Open(filepath.Join("/", "etc", "code", path)).
+			Return(repoMock, nil)
+
+		repoMock.EXPECT().
+			RawCmd([]string{"pull", "--rebase"}).
+			Return(fmt.Sprintf("out: %s", path), nil)
+	}
+
+	if err := codebase.BulkGIT([]string{"pull", "--rebase"}, ch); err != nil {
+		t.Fail()
+	}
+
+	wg.Wait()
+
+	val := sb.String()
+	if !strings.Contains(val, "out: test/something-1") {
+		t.Fail()
+	}
+	if !strings.Contains(val, "out: test/something-2") {
 		t.Fail()
 	}
 }