+++ internal/codebase/provider.go
@@ -31,11 +31,17 @@ const (
 	manifestFile = "manifest.json"
 )
 
+// ProjectEntry map a codebase project entry (i.e the project alongside his codebase local path)
+type ProjectEntry struct {
+	Path    string
+	Project manifest.Project
+}
+
 // Provider is something that allows to Init, Open, or Clone a Codebase
 type Provider interface {
 	Init(path, remote string) (Codebase, error)
 	Open(path string) (Codebase, error)
-	Clone(url, path string) (Codebase, error)
+	Clone(url, path string, ch chan<- ProjectEntry) (Codebase, error)
 }
 
 type provider struct {
@@ -73,8 +79,14 @@ func (provider *provider) Init(path, remote string) (Codebase, error) {
 		return nil, fmt.Errorf("error while creating manifest: %w", err)
 	}
 
+	// create the README.md
+	readmeMd := getReadmeMD(remote)
+	if err := ioutil.WriteFile(filepath.Join(path, metaDir, "README.md"), []byte(readmeMd), 0640); err != nil {
+		return nil, fmt.Errorf("error while writing README.md: %w", err)
+	}
+
 	// Create initial commit
-	if err := repo.CommitFiles("Initial commit", manifestFile); err != nil {
+	if err := repo.CommitFiles("Initial commit", manifestFile, "README.md"); err != nil {
 		return nil, err
 	}
 
@@ -139,7 +151,13 @@ func (provider *provider) Open(path string) (Codebase, error) {
 	}, nil
 }
 
-func (provider *provider) Clone(url, path string) (Codebase, error) {
+func (provider *provider) Clone(url, path string, ch chan<- ProjectEntry) (Codebase, error) {
+	defer func() {
+		if ch != nil {
+			close(ch)
+		}
+	}()
+
 	exist, err := codebaseExists(path)
 	if err != nil {
 		return nil, err
@@ -177,6 +195,13 @@ func (provider *provider) Clone(url, path string) (Codebase, error) {
 				return nil, err
 			}
 		}
+
+		if ch != nil {
+			ch <- ProjectEntry{
+				Path:    projectPath,
+				Project: project,
+			}
+		}
 	}
 
 	return codebase, nil
@@ -193,4 +218,22 @@ func codebaseExists(path string) (bool, error) {
 	}
 
 	return false, err
+}
+
+func getReadmeMD(remote string) string {
+	b := strings.Builder{}
+
+	b.WriteString("# dot-srcode\n\n")
+	b.WriteString("This repository contains a custom .srcode configuration.\n")
+	b.WriteString("See [this repository](https://github.com/creekorful/srcode) for more details.\n\n")
+
+	if remote != "" {
+		b.WriteString("## How to use it\n\n")
+		b.WriteString("One can restore this codebase by issuing the following command:\n\n")
+		b.WriteString("```\n")
+		b.WriteString(fmt.Sprintf("$ srcode clone %s code\n", remote))
+		b.WriteString("```\n")
+	}
+
+	return b.String()
 }